added insert to book controller

+++ BookStore.DataAccess/Migrations/BookStoreContextModelSnapshot.cs
@@ -315,8 +315,8 @@ namespace BookStore.DataAccess.Migrations
                         .HasMaxLength(255)
                         .HasColumnType("nvarchar(255)");
 
-                    b.Property<int>("Size")
-                        .HasColumnType("int");
+                    b.Property<double>("Size")
+                        .HasColumnType("float");
 
                     b.HasKey("Id");
